GenerativeAI-LangGraph/src/langgraph/models.py
class Node:
    def __init__(self, name):
        self.name = name
        self.edges = []

    def add_edge(self, edge):
        self.edges.append(edge)


class Edge:
    def __init__(self, node, weight):
        self.node = node
        self.weight = weight


class LangGraph:
    def __init__(self):
        self.nodes = {}

    def add_node(self, name):
        if name not in self.nodes:
            self.nodes[name] = Node(name)

    def add_edge(self, from_name, to_name, weight):
        if from_name not in self.nodes:
            self.add_node(from_name)
        if to_name not in self.nodes:
            self.add_node(to_name)
        self.nodes[from_name].add_edge(Edge(self.nodes[to_name], weight))

    def process_loan(self, amount: float, term: int, interest_rate: float):
        # Create nodes for loan processing steps
        self.add_node("Application")
        self.add_node("Evaluation")
        self.add_node("Decision")
        self.add_node("Funding")

        # Add edges to represent the loan processing flow
        self.add_edge("Application", "Evaluation", 1)
        self.add_edge("Evaluation", "Decision", 1)
        self.add_edge("Decision", "Funding", 1)

        # You can add more complex logic here to process the loan
        # For example, you could add conditional edges based on the loan details
        if amount > 100000:
            self.add_node("High-Value Review")
            self.add_edge("Evaluation", "High-Value Review", 2)
            self.add_edge("High-Value Review", "Decision", 1)

        # Return some indication of the graph structure
        return f"Loan processing graph created for ${amount} over {term} months at {interest_rate*100}% interest"